+++ changelog.md
@@ -5,6 +5,8 @@
 - Glowing buttons work again with ProjectAzilroka, after Ulduar patch
 - Mage's Heating Up works again with ProjectAzilroka, after Ulduar patch
 - Other features may also work back to normal for ProjectAzilroka users
+- These fixes also apply to addons embedding an outdated LibButtonGlow
+- If there is no fallback solution, glowing buttons are temporarily disabled
 
 #### v0.8.2 (2022-01-18)